requirements.txt
psycopg2-binary
pandas
tqdm
streamlit
openai
python-dotenv 
emoji